% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clues.r
\name{construir_clausula_in}
\alias{construir_clausula_in}
\title{Construir cláusula IN para SQL}
\usage{
construir_clausula_in(clues_seleccionada, clues_info)
}
\arguments{
\item{clues_seleccionada}{Código CLUES seleccionado}

\item{clues_info}{Data frame con información de CLUES}
}
\value{
String con formato SQL para cláusula IN
}
\description{
Construir cláusula IN para SQL
}
