% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_comunes.R
\docType{data}
\name{fecha_corte}
\alias{fecha_corte}
\title{comunes}
\format{
An object of class \code{Date} of length 1.
}
\usage{
fecha_corte
}
\value{
The return value, if any, from executing the utility.
funcion de fecha de corte
}
\description{
A utils function
}
\keyword{datasets}
