% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clues_query.R
\name{mod_clues_query_server}
\alias{mod_clues_query_server}
\title{Server del módulo de consulta CLUES}
\usage{
mod_clues_query_server(id, con, clues_info)
}
\arguments{
\item{id}{ID del módulo}

\item{con}{Conexión a DuckDB (reactiva)}

\item{clues_info}{Data frame con información de CLUES}
}
\description{
Server del módulo de consulta CLUES
}
