% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clues_query.R
\name{mod_clues_query_ui}
\alias{mod_clues_query_ui}
\title{Módulo para consultas de CLUES con DuckDB}
\usage{
mod_clues_query_ui(id)
}
\arguments{
\item{id}{ID del módulo}
}
\description{
Maneja la selección de CLUES y ejecuta consultas incluyendo
las CLUES relacionadas (SSA)
UI del módulo de consulta CLUES
}
