% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clues.r
\name{obtener_clues_relacionadas}
\alias{obtener_clues_relacionadas}
\title{Funciones auxiliares para manejo de CLUES}
\usage{
obtener_clues_relacionadas(clues_seleccionada, clues_info)
}
\arguments{
\item{clues_seleccionada}{Código CLUES seleccionado}

\item{clues_info}{Data frame con información de CLUES (de datos internos)}
}
\value{
Vector con la(s) CLUES a consultar
}
\description{
Funciones para obtener CLUES relacionadas y construir consultas SQL
Obtener CLUES relacionadas (SSA)
}
